End of section 18

+++ client/src/app/messages/messages.component.ts
@@ -1,8 +1,8 @@
-import { MessageService } from './../_services/message.service';
-import { MembersService } from 'src/app/_services/members.service';
-import { Pagination } from '../_models/pagination';
 import { Component, OnInit } from '@angular/core';
 import { Message } from '../_models/message';
+import { Pagination } from '../_models/pagination';
+import { ConfirmService } from '../_services/confirm.service';
+import { MessageService } from '../_services/message.service';
 
 @Component({
   selector: 'app-messages',
@@ -10,36 +10,41 @@ import { Message } from '../_models/message';
   styleUrls: ['./messages.component.css']
 })
 export class MessagesComponent implements OnInit {
-  messages:Message[];
-  pagination : Pagination;
-  container='Unread';
-  pageNumber=1;
-  pageSize=5;
-  loading=false;
+  messages: Message[] = [];
+  pagination: Pagination;
+  container = 'Unread';
+  pageNumber = 1;
+  pageSize = 5;
+  loading = false;
 
-  constructor(private messageService : MessageService) { }
+  constructor(private messageService: MessageService, private confirmService: ConfirmService) { }
 
   ngOnInit(): void {
     this.loadMessages();
   }
 
-  loadMessages(){
-    this.loading=true;
-    this.messageService.getMessages(this.pageNumber, this.pageSize,this.container).subscribe(response =>{
-      this.messages=response.result;
-      this.pagination=response.pagination;
-      this.loading=false;
+  loadMessages() {
+    this.loading = true;
+    this.messageService.getMessages(this.pageNumber, this.pageSize, this.container).subscribe(response => {
+      this.messages = response.result;
+      this.pagination = response.pagination;
+      this.loading = false;
     })
   }
 
-  deleteMessage(id:number) {
-    this.messageService.deleteMessage(id).subscribe(() => {
-      this.messages.splice(this.messages.findIndex(m => m.id === id), 1);
+  deleteMessage(id: number) {
+    this.confirmService.confirm('Confirm delete message', 'This cannot be undone').subscribe(result => {
+      if (result) {
+        this.messageService.deleteMessage(id).subscribe(() => {
+          this.messages.splice(this.messages.findIndex(m => m.id === id), 1);
+        })
+      }
     })
+
   }
 
-  pageChanged(event:any){
-    this.pageNumber=event.page;
+  pageChanged(event: any) {
+    this.pageNumber = event.page;
     this.loadMessages();
   }
